Rename some classes and methods
