+++ flow/src/org/labkey/flow/controllers/protocol/EditICSMetadataForm.java
@@ -15,7 +15,7 @@
  */
 package org.labkey.flow.controllers.protocol;
 
-import org.labkey.api.exp.api.ExpSampleSet;
+import org.labkey.api.exp.api.ExpSampleType;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.ColumnInfo;
@@ -228,7 +228,7 @@ public class EditICSMetadataForm extends ProtocolForm
         }
 
         FieldKey sampleProperty = FieldKey.fromParts("FCSFile", "Sample");
-        ExpSampleSet sampleSet = getProtocol().getSampleSet();
+        ExpSampleType sampleSet = getProtocol().getSampleSet();
         if (sampleSet != null)
         {
             if (sampleSet.hasNameAsIdCol())
