+++ flow/src/org/labkey/flow/controllers/protocol/showProtocol.jsp
@@ -38,7 +38,7 @@
         <labkey:link href="<%=protocol.getSampleSet().detailsURL()%>" text="Show sample type"/><br>
         <labkey:link href="<%=protocol.urlShowSamples()%>" text="Show samples joined to FCS Files" /><br>
         <labkey:link href="<%=protocol.urlUploadSamples()%>" text="Upload more samples from a spreadsheet" /><br>
-        <% if (protocol.getSampleSetJoinFields().size() != 0) { %>
+        <% if (protocol.getSampleTypeJoinFields().size() != 0) { %>
             <labkey:link href="<%=protocol.urlFor(JoinSampleSetAction.class)%>" text="Modify sample join fields" /><br>
         <% } else { %>
             <labkey:link href="<%=protocol.urlFor(JoinSampleSetAction.class)%>" text="Join samples to FCS File Data" /><br>
