+++ microarray/src/org/labkey/microarray/matrix/ExpressionMatrixAssayProvider.java
@@ -166,7 +166,7 @@ public class ExpressionMatrixAssayProvider extends AbstractAssayProvider
     {
         List<NavTree> result = super.getHeaderLinks(viewContext, protocol, containerFilter);
         result.add(new NavTree("manage feature annotation sets", new ActionURL(FeatureAnnotationSetController.ManageAction.class, viewContext.getContainer())));
-        result.add(new NavTree("manage samples", PageFlowUtil.urlProvider(ExperimentUrls.class).getShowSampleSetListURL(viewContext.getContainer())));
+        result.add(new NavTree("manage samples", PageFlowUtil.urlProvider(ExperimentUrls.class).getShowSampleTypeListURL(viewContext.getContainer())));
         return result;
     }
